refactor(celebration): share the confetti burst and add the celebration art
The radial confetti burst was private to the exercise-success panel. Extract
it as-is to a shared component so the Season Pass celebration can reuse the
same motion vocabulary instead of minting a second one.

Ships the celebration panel art as a png/webp/avif triplet (2.0M source
quantized to 748K png / 156K webp / 28K avif).

Suite: 4865 passing.

Wolfcito 🐾 @akawolfcito
